+++ src/main/kotlin/ru/neexol/parsers/excel/Lesson.kt
@@ -1,4 +1,4 @@
-package ru.neexol.models
+package ru.neexol.parsers.excel
 
 class Lesson(
     val name: String,
