+++ src/main/kotlin/ru/neexol/parsers/website/ScheduleFile.kt
@@ -1,4 +1,4 @@
-package ru.neexol.utils
+package ru.neexol.parsers.website
 
 import org.apache.commons.codec.digest.DigestUtils
 import java.net.URL
