+++ src/main/kotlin/ru/neexol/parsers/website/WebsiteParser.kt
@@ -1,4 +1,4 @@
-package ru.neexol.parsers
+package ru.neexol.parsers.website
 
 import org.jsoup.Jsoup
 import java.net.URL
@@ -9,15 +9,15 @@ object WebsiteParser {
             "li:nth-child(1) a.uk-link-toggle, " +
             "li:nth-child(2) a.uk-link-toggle"
 
+    private fun isExamOrTest(s: String) = s.contains("экз") || s.contains("зач")
+    private fun isNotExamOrTest(s: String) = !isExamOrTest(s)
+
+    fun parseLessonsFilesURLs() = parseFilesPaths(::isNotExamOrTest)
+    fun parseExamsOrTestFilesURLs() = parseFilesPaths(::isExamOrTest)
+
     private fun parseFilesPaths(predicate: (String) -> Boolean) = Jsoup.connect(SCHEDULE_URL).get()
         .select(LINKS_SELECTOR)
         .eachAttr("href")
-        .mapNotNull {
-            it.takeIf(predicate)?.let { path ->
-                URL(path)
-            }
-        }
-
-    fun parseLessonsFilesURLs() = parseFilesPaths { !it.contains("экз") }
-    fun parseExamsFilesURLs() = parseFilesPaths { it.contains("экз") }
+        .filter(predicate)
+        .map { ScheduleFile(URL(it)) }
 }
